Rename cli to program so that a cli package can be ripped out of it
