+++ main.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"os"
 
-	"github.com/anttiharju/relcheck/internal/cli"
 	"github.com/anttiharju/relcheck/internal/exitcode"
+	"github.com/anttiharju/relcheck/internal/program"
 	"github.com/anttiharju/relcheck/pkg/interrupt"
 )
 
@@ -13,6 +13,6 @@ func main() {
 	go interrupt.Listen(exitcode.Interrupt, os.Interrupt)
 
 	ctx := context.Background()
-	exitCode := cli.Start(ctx, os.Args[1:])
+	exitCode := program.Start(ctx, os.Args[1:])
 	os.Exit(exitCode)
 }
